// src/models/DebtsModel.ts
export interface DebtModel {
  ano: number
  empenhado: number
  liquidado: number
  pago: number
}
export interface EmpenhadoModel {
  ano: number
  empenhado: number
}
export interface LiquidadoModel {
  ano: number
  liquidado: number
}
export interface PagoModel {
  ano: number
  pago: number
}